data/ukrainian/ua-gec/requirements.txt
ua-gec==2.1.3
pyyaml==6.0.1
